arch-gcn3: convert vALU instruction counters from 32 to 64-bit
The vALU instruction counters were previously 32 bits, but for some
workloads this value wraps around and triggers an assert failure
because the max vALU operations are reached.  To resolve this, this
commit increases the counter size to 64 bits.

Reviewed-on: https://gem5-review.googlesource.com/c/public/gem5/+/29950
Maintainer: Anthony Gutierrez <anthony.gutierrez@amd.com>
Tested-by: kokoro <noreply+kokoro@google.com>

+++ src/gpu-compute/shader.hh
@@ -258,8 +258,8 @@ class Shader : public ClockedObject
     Stats::Vector vectorInstDstOperand;
     void regStats();
 
-    int max_valu_insts;
-    int total_valu_insts;
+    int64_t max_valu_insts;
+    int64_t total_valu_insts;
 
     Shader(const Params *p);
     ~Shader();
